feat(studio-server): /vst/start and /vst/status routes with sidecar adapter
Adds registerVstRoutes (POST /vst/start, GET /vst/status) and the optional
startVstSidecar/getVstSidecarStatus StudioApiAdapter methods, wired into
both hosts (CLI embedded server + Studio Vite dev server).

Relocates the VST sidecar lifecycle module from packages/cli/src/vst/sidecar.ts
to packages/studio-server/src/vstSidecar.ts (exposed via a new
@hyperframes/studio-server/vst-sidecar subpath export) so both hosts can
import it without duplicating the resolver + spawn logic; the CLI module
becomes a thin re-export so hyperframes preview's existing import keeps
working unchanged.
